+++ src/index.ts
@@ -1,1 +1,2 @@
 export { default as ChatBot } from "./chatbot.js";
+export * from "./client.js";